navigation/utils: Break up utils into utils+state #653
Having the documenation atop a utils module won't make sense.
